Hybrid package
