+++ package.json
@@ -20,7 +20,14 @@
     "node": ">=14"
   },
   "type": "module",
-  "main": "index.js",
+  "main": "index.cjs",
+  "module": "index.js",
+  "exports": {
+    ".": {
+      "require":  "./index.cjs",
+      "default": "./index.js"
+    }
+  },
   "scripts": {
     "test": "xo && c8 ava",
     "coverage": "c8 report --reporter=lcov"
@@ -37,11 +44,11 @@
     "validate"
   ],
   "dependencies": {
-    "ansi-escapes": "^5.0.0",
+    "ansi-escapes": "^4.3.2",
     "chalk": "^4.1.2",
-    "log-symbols": "^5.0.0",
-    "plur": "^5.0.0",
-    "string-width": "^5.0.0"
+    "log-symbols": "^4.1.0",
+    "plur": "^4.0.0",
+    "string-width": "^4.2.3"
   },
   "devDependencies": {
     "ava": "^3.15.0",
